requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
typing_extensions==4.7.1
alembic==1.11.0
cryptography==41.0.3
pytest==7.4.2
requests==2.31.0
redis==4.8.2
rq==1.12.0
boto3==1.28.88
python-multipart==0.0.6
PyJWT==2.8.0
psycopg2-binary==2.9.11


